refactor(assistant): move llm-request-log and usage stores into persistence/
